refactor: refactor dbschema
