+++ Makefile
@@ -16,7 +16,7 @@ lint: ## lint
 test: ## test
 	go clean -testcache
 	sh spanner/tools/integration-test.sh
-	sh sqlite/tools/integration-test.sh
+	sh sqlite3/tools/integration-test.sh
 	go clean -testcache
 	go test ./...
 	go test -cover ./... -coverprofile=cover/cover.out && go tool cover -html=cover/cover.out -o cover/cover.html
