--- cli/dbdelete/dbdelete.go
@@ -1,1 +1,0 @@
-package dbdelete
