+++ cmd/dbdelete/README.md
@@ -1,0 +1,41 @@
+# gf-dbdelete
+
+gf-dbdelete is a command-line tool designed to provide a standardized interface for deleting existing rows from tables in a data source. By utilizing a JSON-based input format and providing consistent functionality across data sources.
+
+## Usage
+
+```sh
+gf-dbdelete [-schema <schema-json>] <driver> <data-source>
+gf-dbdelete -h | --help
+```
+
+gf-dbdelete is a command-line tool used to initialize tables in a data source. It supports the following drivers:
+
+- Spanner
+- SQLite3
+
+To use gf-dbdelete with Spanner, specify `spanner` as the `<driver>` and provide a string in the format `projects/<project>/instances/<instance>/databases/<database>` as the `<data-source>`. In this format, `<project>` represents the name of your Google Cloud Platform (GCP) project, `<instance>` is the name of your Spanner instance in the GCP project, and `<database>` is the name of the database within the Spanner instance.
+
+To use gf-dbdelete with SQLite3, set `sqlite3` as the `<driver>` and provide a connection string as the `<data-source>`.
+The connection string should follow the format described in [https://github.com/mattn/go-sqlite3#connection-string](https://github.com/mattn/go-sqlite3#connection-string), such as `file:test.db?cache=shared&mode=memory`.
+
+gf-dbdelete fetches schema information to prepare for the deletion. To provide specific JSON-based schema file, specify the path `<schema-json>` of the file using `-schema` option. The file specified by this option must match the structure described in the Output section in README.md of [dbschema](../dbschema/README.md).
+
+## Input
+
+gf-dbdelete expects a JSON array as input from stdin. The JSON array should have the following structure `DBDeleteInput`:
+
+```ts
+type DBDeleteInput = string[] // list of the table names to be deleted
+```
+
+Here's an example:
+```json
+[ "User", "Comment" ]
+```
+
+In this example, the JSON input instructs gf-dbdelete to delete existing rows in the User and Comment tables.
+
+## Output
+
+There is no specific output generated by gf-dbdelete.
