+++ cmd/dbinit/main.go
@@ -1,18 +1,6 @@
 package main
 
 import (
-	"context"
-	"flag"
-	"fmt"
-	"io"
-	"log"
-	"os"
-
-	"github.com/Jumpaku/gotaface/cli"
-	"github.com/Jumpaku/gotaface/ddl/schema"
-	dbinit_spanner "github.com/Jumpaku/gotaface/spanner/cli/dbinit"
-	dbinit_sqlite "github.com/Jumpaku/gotaface/sqlite/cli/dbinit"
-
 	_ "embed"
 )
 
@@ -20,48 +8,50 @@ import (
 var Usage string
 
 func main() {
-	cmd := flag.NewFlagSet("gf-dbinit", flag.ExitOnError)
-	cmd.Usage = func() { fmt.Println(Usage) }
-	schemaJSON := cmd.String(`schema`, "", "specifies a path <schema-json> of a JSON-based schema file")
+	/*
+		cmd := flag.NewFlagSet("gf-dbinit", flag.ExitOnError)
+		cmd.Usage = func() { fmt.Println(Usage) }
+		schemaJSON := cmd.String(`schema`, "", "specifies a path <schema-json> of a JSON-based schema file")
 
-	if err := cmd.Parse(os.Args[1:]); err != nil {
-		log.Fatalf(`cannot parse command line arguments: %v`, err)
-	}
-
-	args := cmd.Args()
-	if len(args) != 2 {
-		log.Fatalln(`positional arguments <driver> and <data-source> are required`)
-	}
-
-	driver, dataSource := args[0], args[1]
-	var fetcher schema.Fetcher
-	if schemaJSON := *schemaJSON; schemaJSON != "" {
-		f, err := os.Open(schemaJSON)
-		if err != nil {
-			log.Fatalf(`fail to open file: %s: %v`, schemaJSON, err)
+		if err := cmd.Parse(os.Args[1:]); err != nil {
+			log.Fatalf(`cannot parse command line arguments: %v`, err)
 		}
 
-		b, err := io.ReadAll(f)
-		if err != nil {
-			log.Fatalf(`fail to read file: %s : %v`, schemaJSON, err)
+		args := cmd.Args()
+		if len(args) != 2 {
+			log.Fatalln(`positional arguments <driver> and <data-source> are required`)
 		}
 
-		fetcher = schema.NewFetcher(b)
-	}
+		driver, dataSource := args[0], args[1]
+		var fetcher schema.Fetcher
+		if schemaJSON := *schemaJSON; schemaJSON != "" {
+			f, err := os.Open(schemaJSON)
+			if err != nil {
+				log.Fatalf(`fail to open file: %s: %v`, schemaJSON, err)
+			}
 
-	var runner cli.Runner
-	switch driver {
-	default:
-		log.Fatalf(`unsupported driver %s`, driver)
-	case `spanner`:
-		runner = &dbinit_spanner.SpannerRunner{DataSource: dataSource, Fetcher: fetcher}
-	case `sqlite3`:
-		runner = &dbinit_sqlite.SqliteRunner{DataSource: dataSource, Fetcher: fetcher}
-	}
+			b, err := io.ReadAll(f)
+			if err != nil {
+				log.Fatalf(`fail to read file: %s : %v`, schemaJSON, err)
+			}
 
-	ctx := context.Background()
-	err := runner.Run(ctx, os.Stdin, os.Stdout)
-	if err != nil {
-		log.Fatalf(`failed execution: %v`, err)
-	}
+			fetcher = schema.NewFetcher(b)
+		}
+
+		var runner cli.Runner
+		switch driver {
+		default:
+			log.Fatalf(`unsupported driver %s`, driver)
+		case `spanner`:
+			//runner = &dbinit_spanner.SpannerRunner{DataSource: dataSource, Fetcher: fetcher}
+		case `sqlite3`:
+			//runner = &dbinit_sqlite.SqliteRunner{DataSource: dataSource, Fetcher: fetcher}
+		}
+
+		ctx := context.Background()
+		err := runner.Run(ctx, os.Stdin, os.Stdout)
+		if err != nil {
+			log.Fatalf(`failed execution: %v`, err)
+		}
+	*/
 }
