--- sqlite/tools/integration-test.sh
@@ -1,6 +1,0 @@
-#!/bin/sh
-
-export ENV_SQLITE_TEST_DIR=$(pwd)/sqlite/test_db
-
-mkdir -p "${ENV_SQLITE_TEST_DIR}"
-go test ./sqlite/cli/dbschema/...
