+++ sqlite3/cli/dbdelete/runner.go
@@ -8,9 +8,9 @@ import (
 
 	"github.com/Jumpaku/gotaface/cli/dbinit"
 	"github.com/Jumpaku/gotaface/ddl/schema"
-	sqlite_schema "github.com/Jumpaku/gotaface/sqlite/ddl/schema"
-	sqlite_delete "github.com/Jumpaku/gotaface/sqlite/dml/delete"
-	sqlite_insert "github.com/Jumpaku/gotaface/sqlite/dml/insert"
+	sqlite3_schema "github.com/Jumpaku/gotaface/sqlite3/ddl/schema"
+	sqlite3_delete "github.com/Jumpaku/gotaface/sqlite3/dml/delete"
+	sqlite3_insert "github.com/Jumpaku/gotaface/sqlite3/dml/insert"
 )
 
 type SqliteRunner struct {
@@ -24,13 +24,13 @@ func (r *SqliteRunner) Run(ctx context.Context, stdin io.Reader, stdout io.Write
 		return fmt.Errorf(`fail to load table initialization data from stdin: %w`, err)
 	}
 
-	db, err := sql.Open("sqlite", r.DataSource)
+	db, err := sql.Open("sqlite3", r.DataSource)
 	if err != nil {
-		return fmt.Errorf(`fail to create sqlite client: %w`, err)
+		return fmt.Errorf(`fail to create sqlite3 client: %w`, err)
 	}
 	var fetcher = r.Fetcher
 	if fetcher == nil {
-		fetcher = sqlite_schema.NewFetcher(db)
+		fetcher = sqlite3_schema.NewFetcher(db)
 	}
 	schema, err := fetcher.Fetch(ctx)
 	if err != nil {
@@ -42,12 +42,12 @@ func (r *SqliteRunner) Run(ctx context.Context, stdin io.Reader, stdout io.Write
 		return fmt.Errorf(`fail to prepare tables: %w`, err)
 	}
 
-	err = dbinit.DeleteRowsInParallel(ctx, sqlite_delete.NewDeleter(db), deleteTables)
+	err = dbinit.DeleteRowsInParallel(ctx, sqlite3_delete.NewDeleter(db), deleteTables)
 	if err != nil {
 		return fmt.Errorf(`fail to delete rows in tables: %w`, err)
 	}
 
-	err = dbinit.InsertRowsInParallel(ctx, sqlite_insert.NewInserter(db), insertTableRows)
+	err = dbinit.InsertRowsInParallel(ctx, sqlite3_insert.NewInserter(db), insertTableRows)
 	if err != nil {
 		return fmt.Errorf(`fail to insert rows in tables: %w`, err)
 	}
