+++ sqlite3/test/env.go
